+++ CHANGELOG.md
@@ -3,9 +3,10 @@ History of user-visible changes.
 ## [v1.5.3](https://github.com/openconnect/openconnect-gui/tree/v1.5.3) (unreleased)
 - No disconnection triggered before quit (#202)
 - Don't use system wide defined proxy when disabled in profile (#208)
-- update Qt (5.10.0)
 - Unable to use socks5 proxy built by ssh tunneling (#187)
-- update spdlog (0.16.3)
+
+- update Qt(5.10.1), spdlog(0.16.3)
+- update GnuTLS(3.6.1), p11-kit(0.23.9), stoken(0.92), libproxy(0.4.15), nettle(3.4), lz4(1.8.1.2), libxml2(2.9.7), libidn2(2.0.4)
 
 [Full Changelog](https://github.com/openconnect/openconnect-gui/compare/v1.5.2...v1.5.3)